+++ demo/src/app/components/+empty/empty-section.component.ts
@@ -1,0 +1,16 @@
+
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'demo-empty-section',
+    templateUrl: './empty-section.component.html',
+    styleUrls: [
+        './empty-section.scss'
+    ]
+})
+export class DemoEmptySectionComponent {
+
+    constructor() {
+    }
+
+}