+++ src/empty/empty.config.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class ThyEmptyConfig {
+    noResultTranslateId = 'common.tips.NO_RESULT';
+    noResultWithTargetTranslateId = 'common.tips.NO_RESULT_TARGET';
+}
+