+++ src/index.ts
@@ -1,72 +1,8 @@
-import { Component, NgModule } from '@angular/core';
-import { ModalModule } from 'ngx-bootstrap/modal';
 
-import { ThySharedModule } from './shared/shared.module';
-import { ThyButtonModule } from './button/button.module';
-import { ThyLayoutModule } from './layout/layout.module';
-import { NgxPopBoxModule } from './pop-box/pop-box.module';
-import { ThyGridModule } from './grid/grid.module';
-import { ThyAvatarModule } from './avatar/avatar.module';
-import { ThyBadgeModule } from './badge/badge.module';
-import { ThyLabelModule } from './label/label.module';
-import { ThyNavModule } from './nav/nav.module';
-import { ThyPaginationModule } from './pagination/pagination.module';
-import { ThyModalModule } from './modal/modal.module';
-import { ThyCardModule } from './card/card.module';
-import { ThyLoadingModule } from './loading/loading.module';
-import { ThyActionMenuModule } from './action-menu/action-menu.module';
-import { ThyConfirmModule } from './confirm/confirm.module';
-import { ThyTreeModule } from './tree/tree.module';
-
-
-@NgModule({
-    declarations: [
-    ],
-    imports: [
-        ModalModule.forRoot(),
-        ThyLayoutModule,
-        ThyButtonModule,
-        NgxPopBoxModule,
-        ThyBadgeModule,
-        ThyGridModule,
-        ThyAvatarModule,
-        ThyLabelModule,
-        ThyNavModule,
-        ThyPaginationModule,
-        ThyModalModule,
-        ThyCardModule,
-        ThyLoadingModule,
-        ThyActionMenuModule,
-        ThyConfirmModule,
-    ],
-    exports: [
-        ThyLayoutModule,
-        ThyButtonModule,
-        ThyBadgeModule,
-        ThyGridModule,
-        ThyAvatarModule,
-        ThyLabelModule,
-        ThyNavModule,
-        ThyPaginationModule,
-        ThyModalModule,
-        ThyCardModule,
-        ThyLoadingModule,
-        ThyActionMenuModule,
-        ThyConfirmModule,
-        ThyTreeModule
-    ],
-    providers: [
-    ]
-})
-export class NgxTethysModule {
-
-}
-
-
-export {
-    ThyPopBoxService
-} from './pop-box';
-
-export {
-    ThyModalService
-} from './modal';
+export * from './module';
+export { ThyPopBoxService } from './pop-box';
+export { ThyModalService } from './modal';
+export { ThyNotifyService } from './notify/notify.service';
+export { ThyConfirmService } from './confirm/confirm.service';
+export { ThyTranslate } from './shared';
+export * from './store';