+++ src/store/index.ts
@@ -1,0 +1,3 @@
+export * from './module';
+export * from './store';
+export * from './action';