+++ src/store/module.ts
@@ -1,0 +1,23 @@
+import { NgModule, ModuleWithProviders } from '@angular/core';
+import { ROOT_STATE_TOKEN } from './types';
+
+@NgModule()
+export class RootStoreModule {
+
+}
+@NgModule({})
+export class ThyStoreModule {
+
+    static forRoot(stores: any[] = []): ModuleWithProviders {
+        return {
+            ngModule: RootStoreModule,
+            providers: [
+                ...stores,
+                {
+                    provide: ROOT_STATE_TOKEN,
+                    useValue: stores
+                }
+            ]
+        };
+    }
+}