+++ src/store/util.ts
@@ -1,0 +1,14 @@
+import { StoreMetaInfo, META_KEY } from './types';
+
+export function findAndCreateStoreMetadata(target: any): StoreMetaInfo {
+    if (!target.hasOwnProperty(META_KEY)) {
+        const defaultMetadata: StoreMetaInfo = {
+            actions: {},
+            path: null,
+            children: [],
+            instance: null
+        };
+        target[META_KEY] = defaultMetadata;
+    }
+    return target[META_KEY];
+}