+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-## 0.0.14
+## 0.0.21
+export RootStoreModule
+
+## 0.0.20
+fix store build error.
+
+## 0.0.19
+fix ts define file error.
+
+## 0.0.18
+1. add thy-empty component;
+1. add thy store for state management.
 
 ## 0.0.13